requirements.txt
scipy
numpy
pandas
matplotlib
jax
dm-haiku
optax
pyarrow
fastparquet
seaborn

figrid
imageio
